+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 * `Exit()` is no longer conditionally-built
   - dents our code coverage figures, but more important that it's available when you build your tests
+* `TestFilepathExists()` no longer writes to the pipeline on success
+  - now more accurate for how a UNIX shell `[[ -e filepath ]]` behaves
 
 ## v1.0.0